tools/deepl-translate-en/requirements.txt
requests
pyperclip
python-dotenv